requirements.txt
fastapi
uvicorn
jinja2
python-multipart
telethon
sqlalchemy
pydantic
aiofiles
itsdangerous
requests
crcmod
passlib[argon2]
argon2-cffi
bcrypt
pyotp
python-dotenv